developer/django_practices/api/apispractices/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
django-cors-headers==4.0.0
drf-yasg==1.21.5
pytest-django==4.5.2